added offline dialog cards, deleted unusable resources, bug fixes

+++ app/src/main/java/az/siftoshka/habitube/utils/CurrencyFormatter.java
@@ -4,7 +4,7 @@ import java.text.DecimalFormat;
 
 public class CurrencyFormatter {
 
-    public static String format(int number) {
+    public static String format(long number) {
         DecimalFormat decimalFormat = new DecimalFormat("###,###,##0");
         return decimalFormat.format(Double.parseDouble(String.valueOf(number)));
     }
